requirements.txt
uagents
openai
python-dotenv
langchain
langchain-openai
deepagents
